Air Cargo
SQL Project | Air Cargo is an aviation company that provides air transportation services for passengers and freight. Air Cargo
uses its aircraft to provide different services with the help of partnerships or alliances with other airlines. The company wants to
prepare reports on regular passengers, busiest routes, ticket sales details, and other scenarios to improve the ease of travel and
booking for customers.