ScienceQtech
SQL Project | ScienceQtech is a startup that works in the Data Science field. It has worked on fraud detection, market basket, selfdriving
cars, supply chain, algorithmic early detection of lung cancer, and the drug discovery field. The HR department wants you
to generate reports on employee details, their performance, and on the project that the employees have undertaken, to analyze the
employee database and extract specific data based on different requirements.